Consolidate notebook metadata and realtime offer handshakes
