+++ .changeset/codex-shared-metadata-and-voice.md
@@ -1,0 +1,5 @@
+---
+"@howaboua/pi-codex-conversion": patch
+---
+
+Simplify Notebook metadata persistence and realtime voice startup without changing saved data, connection behavior, or tool output.
