+++ packages/pi-codex-conversion/src/voice/lan/browser-peer.ts
@@ -1,8 +1,7 @@
 import type { CodexConversionConfig } from "../../adapter/activation/config.ts";
 import { VoiceHelperClient, type VoiceHelperEvent } from "../helper.ts";
+import { startRealtimeOffer } from "../conversation/helper-offer.ts";
 import type { CodexRealtimePeerEvent, CodexRealtimeWebRtcPeer } from "../conversation/peer.ts";
-
-const OFFER_TIMEOUT_MS = 15_000;
 
 export class LanHostRealtimePeer implements CodexRealtimeWebRtcPeer {
 	readonly kind = "webrtc" as const;
@@ -28,26 +27,8 @@ export class LanHostRealtimePeer implements CodexRealtimeWebRtcPeer {
 		return this.helper.onExit(listener);
 	}
 
-	async start(config: CodexConversionConfig): Promise<string> {
-		await this.helper.start(config.tools.customRustBinariesDir);
-		if (this.helper.protocolVersion !== 5) {
-			const actualVersion = this.helper.protocolVersion ?? "unknown";
-			await this.helper.close();
-			throw new Error(`Incompatible Codex voice helper protocol ${actualVersion}; expected 5`);
-		}
-		const offer = Promise.withResolvers<string>();
-		const removeEvent = this.helper.onEvent((event) => {
-			if (event.type === "offer") offer.resolve(event.sdp);
-			else if (event.type === "error") offer.reject(new Error(event.message));
-		});
-		const removeExit = this.helper.onExit((error) => offer.reject(error));
-		const timeout = setTimeout(() => offer.reject(new Error("Codex voice helper did not create an offer")), OFFER_TIMEOUT_MS);
-		this.helper.send({ type: "start_v3_bridge" });
-		return offer.promise.finally(() => {
-			clearTimeout(timeout);
-			removeEvent();
-			removeExit();
-		});
+	start(config: CodexConversionConfig): Promise<string> {
+		return startRealtimeOffer(this.helper, config, "bridge");
 	}
 
 	applyAnswer(sdp: string): void {
